requirements.txt
alembic==1.7.3
blinker==1.4
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
onetimepass==1.0.1
passlib==1.7.4
PyJWT==2.3.0
PyMySQL==1.0.2
PyQRCode==1.2.1
python-dotenv==0.19.0
simplejson==3.17.5
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.0.1
